# SMF_FINAL/smf/docs/smf/api-docs.yaml
openapi: 3.0.0
info:
  title: SMF (Security Management Framework) API
  version: 1.0.0
paths:
...
  /api/v1/events/client:
    get:
      summary: Get client events
      operationId: >-
        home_joe-arch_documents_bruno_smf_security_management_framework_api_event-controller_get_client_events_yml
      description: Retrieves events that have occurred since the specified time
      tags:
        - event-controller
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                  time:
                    type: string
              example:
                success: true
                message: Events retrieved successfully
                data:
                  - eventId: 550e8400-e29b-41d4-a716-446655440000
                    deviceId: 550e8400-e29b-41d4-a716-446655440001
                    macAddress: 00:1A:2B:3C:4D:5E
                    event: DEVICE_ONLINE
                    timestamp: '2024-01-15T10:25:00Z'
                time: '2024-01-15T10:30:00Z'
      parameters:
        - name: since
          in: query
          description: Filter events that occurred after this timestamp
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
  /api/v1/events/device:
    post:
      summary: Submit a device event
      operationId: >-
        home_joe-arch_documents_bruno_smf_security_management_framework_api_event-controller_submit_a_device_event_yml
      description: >-
        Receives events from devices such as status changes, access events, and alerts
      tags:
        - event-controller
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                  time:
                    type: string
              example:
                success: true
                message: Event completed successfully
                data: {}
                time: '2024-01-15T10:30:00Z'
      requestBody:
        $ref: '#/components/requestBodies/submit_a_device_event'
  /api/v1/roles/:
    post:
      summary: Create a new role
      operationId: >-
        home_joe-arch_documents_bruno_smf_security_management_framework_api_role-controller_create_a_new_role_yml
      description: Creates a new role with the specified name
      tags:
        - role-controller
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                  time:
                    type: string
              example:
                success: true
                message: Role created successfully
                data:
                  id: 1
                  roleName: MANAGER
                time: '2024-01-15T10:30:00Z'
      requestBody:
        $ref: '#/components/requestBodies/create_a_new_role'
    get:
      summary: Get all roles
      operationId: >-
        home_joe-arch_documents_bruno_smf_security_management_framework_api_role-controller_get_all_roles_yml
      description: Retrieves all roles in the system
      tags:
        - role-controller
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                  time:
                    type: string
              example:
                success: true
                message: Roles retrieved successfully
                data:
                  - id: 1
                    roleName: ADMIN
                  - id: 2
                    roleName: USER
                  - id: 3
                    roleName: OPERATOR
                time: '2024-01-15T10:30:00Z'
  /api/v1/roles/:id:
    delete:
      summary: Delete a role
      operationId: >-
        home_joe-arch_documents_bruno_smf_security_management_framework_api_role-controller_delete_a_role_yml
      description: Deletes a role by its ID
      tags:
        - role-controller
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                  time:
                    type: string
              example:
                success: true
                message: Role deleted successfully
                data: {}
                time: '2024-01-15T10:30:00Z'
        parameters:
        - name: id
          in: path
          description: ID of the role to delete
          required: true
          schema:
            type: string
    get:
      summary: Get a role by ID
      operationId: >-
        home_joe-arch_documents_bruno_smf_security_management_framework_api_role-controller_get_a_role_by_id_yml
      description: Retrieves a specific role by its ID
      tags:
        - role-controller
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                  time:
                    type: string
              example:
                success: true
                message: Role retrieved successfully
                data:
                  id: 1
                  roleName: ADMIN
                time: '2024-01-15T10:30:00Z'
        parameters:
        - name: id
          in: path
          description: ID of the role
          required: true
          schema:
            type: string
    put:
      summary: Update an existing role
      operationId: >-
        home_joe-arch_documents_bruno_smf_security_management_framework_api_role-controller_update_an_existing_role_yml
      description: Updates the name of an existing role
      tags:
        - role-controller
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                  time:
                    type: string
              example:
                success: true
                message: Role updated successfully
                data: {}
                time: '2024-01-15T10:30:00Z'
        parameters:
        - name: id
          in: path
          description: ID of the role to update
          required: true
          schema:
            type: string
      requestBody:
        $ref: '#/components/requestBodies/update_an_existing_role'
...
  /api/v1/zones/:id/roles/:roleId:
    post:
      summary: Assign a role to a zone
      operationId: >-
        home_joe-arch_documents_bruno_smf_security_management_framework_api_zone-controller_assign_a_role_to_a_zone_yml
      description: Associates a role with a specific zone
      tags:
        - zone-controller
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                  time:
                    type: string
              example:
                success: true
                message: Role assigned to zone successfully
                data: {}
                time: '2024-01-15T10:30:00Z'
        parameters:
        - name: id
          in: path
          description: UUID of the zone
          required: true
          schema:
            type: string
        - name: roleId
          in: path
          description: ID of the role to assign
          required: true
          schema:
            type: string
    delete:
      summary: Remove a role from a zone
      operationId: >-
        home_joe-arch_documents_bruno_smf_security_management_framework_api_zone-controller_remove_a_role_from_a_zone_yml
      description: Removes a role association from a specific zone
      tags:
        - zone-controller
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                  time:
                    type: string
              example:
                success: true
                message: Role removed from zone successfully
                data: {}
                time: '2024-01-15T10:30:00Z'
        parameters:
        - name: id
          in: path
          description: UUID of the zone
          required: true
          schema:
            type: string
        - name: roleId
          in: path
          description: ID of the role to remove
          required: true
          schema:
            type: string
...
  /api/v1/workers/:id:
    get:
      summary: Get a worker by ID
      operationId: get_worker
      description: Retrieves the full worker profile from Supabase by UUID. Requires ADMIN role.
      tags:
        - worker-controller
      parameters:
        - name: id
          in: path
          description: UUID of the worker
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/worker_response'
                  time:
                    type: string
              example:
                success: true
                message: Worker fetched successfully
                data:
                  id: a1b2c3d4-e5f6-7890-abcd-ef1234567890
                  full_name_ar: عمر حسن الرشيدي
                  full_name_en: Omar Hassan Al-Rashidi
                  date_of_birth: '1985-03-14'
                  address_en: King Fahd Road, Riyadh
                  phone: '0501234567'
                  role_en: Construction Worker
                  company_en: United Construction Co.
                  work_location_en: City Towers Project
                  medical_condition_en: Type 2 Diabetes
                  clinical_notes_en: Requires regular glucose monitoring and oral medications.
                  emergency_contact_name: Sarah Omar
                  emergency_contact_relation: Wife
                  emergency_phone: '0509876543'
                  created_at: '2026-05-26T12:00:00Z'
                  updated_at: '2026-05-26T12:00:00Z'
                time: '2026-05-26T12:00:00Z'
        '404':
          description: Not Found
      security:
        - bearerAuth: []

    patch:
      summary: Update a worker
      operationId: update_worker
      description: >-
        Updates an existing worker record in Supabase. All fields are optional —
        only send the fields you want to change. Requires ADMIN role.
      tags:
        - worker-controller
      parameters:
        - name: id
          in: path
          description: UUID of the worker to update
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        $ref: '#/components/requestBodies/update_worker'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/worker_response'
                  time:
                    type: string
              example:
                success: true
                message: Worker updated successfully
                data:
                  id: a1b2c3d4-e5f6-7890-abcd-ef1234567890
                  full_name_en: Omar Hassan Al-Rashidi
                  medical_condition_en: Type 2 Diabetes — controlled
                  updated_at: '2026-05-26T14:00:00Z'
                time: '2026-05-26T14:00:00Z'
        '404':
          description: Not Found
      security:
        - bearerAuth: []

    delete:
      summary: Delete a worker
      operationId: delete_worker
      description: Deletes a worker record from Supabase by UUID. Requires ADMIN role.
      tags:
        - worker-controller
      parameters:
        - name: id
          in: path
          description: UUID of the worker to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    nullable: true
                  time:
                    type: string
              example:
                success: true
                message: Worker deleted successfully
                data: null
                time: '2026-05-26T14:00:00Z'
        '404':
          description: Not Found
      security:
        - bearerAuth: []
...
servers:
  - url: http://localhost:8080
    description: Base Server
  - url: http://localhost:8080
    description: Local development server
components:
  schemas:
...
    submit_a_device_event:
      type: object
      properties:
        macAddress:
          type: string
        event:
          type: string
      example:
        macAddress: 00:1A:2B:3C:4D:5E
        event: DEVICE_ONLINE
    create_a_new_role:
      type: object
      properties:
        roleName:
          type: string
      example:
        roleName: MANAGER
    update_an_existing_role:
      type: object
      properties:
        roleName:
          type: string
      example:
        roleName: SUPER_ADMIN
...
    worker_request:
      type: object
      required:
        - full_name_en
      properties:
        full_name_ar:
          type: string
          description: Full name in Arabic
        full_name_en:
          type: string
          description: Full name in English (required)
        date_of_birth:
          type: string
          format: date
        address_ar:
          type: string
        address_en:
          type: string
        phone:
          type: string
        role_ar:
          type: string
        role_en:
          type: string
        company_ar:
          type: string
        company_en:
          type: string
        work_location_ar:
          type: string
        work_location_en:
          type: string
        medical_condition_ar:
          type: string
        medical_condition_en:
          type: string
        clinical_notes_ar:
          type: string
        clinical_notes_en:
          type: string
        emergency_contact_name:
          type: string
        emergency_contact_relation:
          type: string
        emergency_phone:
          type: string
      example:
        full_name_ar: عمر حسن الرشيدي
        full_name_en: Omar Hassan Al-Rashidi
        date_of_birth: '1985-03-14'
        address_ar: شارع الملك فهد، الرياض
        address_en: King Fahd Road, Riyadh
        phone: '0501234567'
        role_ar: عامل بناء
        role_en: Construction Worker
        company_ar: شركة الإنشاءات المتحدة
        company_en: United Construction Co.
        work_location_ar: مشروع أبراج المدينة
        work_location_en: City Towers Project
        medical_condition_ar: داء السكري من النوع الثاني
        medical_condition_en: Type 2 Diabetes
        clinical_notes_ar: يحتاج إلى مراقبة الجلوكوز بانتظام وتناول الأدوية عن طريق الفم.
        clinical_notes_en: Requires regular glucose monitoring and oral medications.
        emergency_contact_name: Sarah Omar
        emergency_contact_relation: Wife
        emergency_phone: '0509876543'
    worker_response:
      type: object
      properties:
        id:
          type: string
          format: uuid
        full_name_ar:
          type: string
        full_name_en:
          type: string
        date_of_birth:
          type: string
          format: date
        address_ar:
          type: string
        address_en:
          type: string
        phone:
          type: string
        role_ar:
          type: string
        role_en:
          type: string
        company_ar:
          type: string
        company_en:
          type: string
        work_location_ar:
          type: string
        work_location_en:
          type: string
        medical_condition_ar:
          type: string
        medical_condition_en:
          type: string
        clinical_notes_ar:
          type: string
        clinical_notes_en:
          type: string
        emergency_contact_name:
          type: string
        emergency_contact_relation:
          type: string
        emergency_phone:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
  requestBodies:
...
    submit_a_device_event:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/submit_a_device_event'
      description: ''
      required: true
    create_a_new_role:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/create_a_new_role'
      description: ''
      required: true
    update_an_existing_role:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/update_an_existing_role'
      description: ''
      required: true
...
    update_worker:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/worker_request'
      description: All fields optional — include only what you want to change.
      required: true
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: JWT token from POST /api/v1/auth/login
    add_device_to_registry:
      type: http
      scheme: bearer
    remove_device_from_registry:
      type: http
      scheme: bearer
    get_smf_device_by_id:
      type: http
      scheme: bearer
    get_smf_device_by_label:
      type: http
      scheme: bearer
    get_all_smf_devices:
      type: http
      scheme: bearer
    get_unregistered_smf_devices:
      type: http
      scheme: bearer
    create_an_announcement:
      type: http
      scheme: bearer
    get_all_announcements:
      type: http
      scheme: bearer
    get_announcement_by_id:
      type: http
      scheme: bearer
    delete_announcement:
      type: http
      scheme: bearer